fix: rename redux folders
