+++ client/src/pages/SignUp/index.tsx
@@ -2,8 +2,8 @@ import { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { Link } from "react-router-dom";
 import { PATH } from "../../consts";
-import { userSlice } from "../../state/user/reducer";
 import { getIsAuthorized } from "../../state/user/selectors";
+import { userSlice } from "../../state/user/slice";
 import { AppDispatch } from "../../store";
 import { FormValues, SignUpForm } from "./SignUpForm";
