+++ client/src/state/user/slice.test.ts
@@ -1,4 +1,4 @@
-import { userSlice, UserState } from "./reducer";
+import { userSlice, UserState } from "./slice";
 
 describe("pizzaReducer", () => {
   describe("set_isAuthorized", () => {
